Add run_train training loop; fix get_batch dtype for uint16 data
Wire up the end-to-end training loop (dataclass configs, AdamW + cosine
LR schedule, grad clipping, checkpointing) over a memmapped uint16 token
array. Cast get_batch tensors to torch.long so uint16 token ids can index
the embedding / serve as cross-entropy targets. Verified: loss drops from
~ln(vocab) (9.2) to ~4.0 on TinyStories valid. Ignore checkpoints/.

+++ cs336_basics/transformer.py
@@ -493,8 +493,8 @@ def get_batch(
     )
     offsets = np.arange(context_length)
     indices = begins[:, None] + offsets[None, :]
-    input_tokens = torch.tensor(dataset[indices], device=device)
-    next_tokens = torch.tensor(dataset[indices + 1], device=device)
+    input_tokens = torch.tensor(dataset[indices], device=device, dtype=torch.long)
+    next_tokens = torch.tensor(dataset[indices + 1], device=device, dtype=torch.long)
     return (input_tokens, next_tokens)
